add tab bar
