+++ src/App.js
@@ -1,65 +1,14 @@
 // @flow
 
 import React from "react";
-import { StyleSheet, Text, SafeAreaView } from "react-native";
-import { getEvents, updateEvents } from "./integrations/cms";
-import type { Event } from "./integrations/cms";
-import EventList from "./EventList";
-import { bgColor } from "./constants/colors";
+import { View } from "react-native";
+import { TabNavigator } from "react-navigation";
+import EventsScreen from "./screens/EventsScreen";
 
-const locale = "en-GB";
-
-type State = {
-  events: Event[],
-  loaded: boolean,
-  refreshing: boolean
-};
-
-class App extends React.Component<{}, State> {
-  constructor() {
-    super();
-
-    this.state = {
-      events: [],
-      loaded: false,
-      refreshing: false
-    };
-  }
-
-  async componentDidMount() {
-    const events = await getEvents();
-
-    // eslint-disable-next-line react/no-did-mount-set-state
-    this.setState({ events, loaded: true });
-  }
-
-  onRefreshEvents = () => {
-    this.setState({ refreshing: true }, async () => {
-      const events = await updateEvents();
-      this.setState({ events, refreshing: false });
-    });
-  };
-
-  render() {
-    return (
-      <SafeAreaView style={styles.container}>
-        {!this.state.loaded && <Text>Loading...</Text>}
-        <EventList
-          locale={locale}
-          events={this.state.events}
-          refreshing={this.state.refreshing}
-          onRefresh={this.onRefreshEvents}
-        />
-      </SafeAreaView>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: bgColor
-  }
+export default TabNavigator({
+  Home: { screen: EventsScreen },
+  Saved: { screen: () => <View /> },
+  Parade: { screen: () => <View /> },
+  Donate: { screen: () => <View /> },
+  More: { screen: () => <View /> }
 });
-
-export default App;
